requirements.txt
Pillow==2.3.0
ipython==1.2.1
python-magic==0.4.6
readline==6.2.4.1
wsgiref==0.1.2
